Move operation class to its own module;
Change tests to use production class;
Add new fixtures for operation;

PWPA-2002
