+++ src/alfasim_score/conftest.py
@@ -3,6 +3,7 @@ from pathlib import Path
 
 from alfasim_score.converter.alfacase.base_operation import BaseOperationBuilder
 from alfasim_score.converter.alfacase.convert_alfacase import ScoreAlfacaseConverter
+from alfasim_score.converter.alfacase.production_operation import ProductionOperationBuilder
 from alfasim_score.converter.alfacase.score_input_reader import ScoreInputReader
 
 
@@ -34,3 +35,13 @@ def base_operation_gas_lift(shared_datadir: Path) -> BaseOperationBuilder:
 @pytest.fixture
 def base_operation_natural_flow(shared_datadir: Path) -> BaseOperationBuilder:
     return BaseOperationBuilder(shared_datadir / "score_input_natural_flow.json")
+
+
+@pytest.fixture
+def production_operation_gas_lift(shared_datadir: Path) -> ProductionOperationBuilder:
+    return ProductionOperationBuilder(shared_datadir / "score_input_gas_lift.json")
+
+
+@pytest.fixture
+def production_operation_natural_flow(shared_datadir: Path) -> ProductionOperationBuilder:
+    return ProductionOperationBuilder(shared_datadir / "score_input_natural_flow.json")
