+++ src/alfasim_score/converter/alfacase/_tests/test_convert_alfacase.py
@@ -5,7 +5,7 @@ from pytest_regressions.file_regression import FileRegressionFixture
 
 from alfasim_score.converter.alfacase.base_operation import BaseOperationBuilder
 from alfasim_score.converter.alfacase.convert_alfacase import ScoreAlfacaseConverter
-from alfasim_score.converter.alfacase.score_input_reader import ScoreInputReader
+from alfasim_score.converter.alfacase.production_operation import ProductionOperationBuilder
 
 
 def test_create_alfacase_gas_lift(
@@ -30,9 +30,9 @@ def test_create_alfacase_natural_flow(
 
 def test_create_alfacase_gas_lift_with_operation(
     file_regression: FileRegressionFixture,
-    base_operation_gas_lift: BaseOperationBuilder,
+    production_operation_gas_lift: ProductionOperationBuilder,
 ) -> None:
-    case_description = base_operation_gas_lift.generate_operation_alfacase_description()
+    case_description = production_operation_gas_lift.generate_operation_alfacase_description()
     file_regression.check(
         convert_description_to_alfacase(case_description), encoding="utf-8", extension=".alfacase"
     )
@@ -40,9 +40,9 @@ def test_create_alfacase_gas_lift_with_operation(
 
 def test_create_alfacase_natural_flow_with_operation(
     file_regression: FileRegressionFixture,
-    base_operation_natural_flow: BaseOperationBuilder,
+    production_operation_natural_flow: ProductionOperationBuilder,
 ) -> None:
-    case_description = base_operation_natural_flow.generate_operation_alfacase_description()
+    case_description = production_operation_natural_flow.generate_operation_alfacase_description()
     file_regression.check(
         convert_description_to_alfacase(case_description), encoding="utf-8", extension=".alfacase"
     )
